Revert "rclient.sh is useless now." + Remove LD_LIBRARY_PATH
This reverts commit c8943e25d62e601d7c3e1b41e5e81803f8f079b1.

Also remove LD_LIBRARY_PATH since it is really useless now.

pid 1 has special handling in either host or in container (e.g.
signal handling), so we'd the wrapper to be pid 1 although this
is not necessary. In addition this seems to make us easier to write
some test cases.